+++ TellarknightApp/Models/Card.cs
@@ -16,5 +16,11 @@ namespace TellarknightApp.Models
         public string? Role { get; set; }
         public string? Image { get; set; }
         public int Quantity { get; set; }
+
+        public virtual void AnalyzeHand(string name)
+        {
+            string result = $"CardClass: {name}";
+            Console.WriteLine(result);
+        }
     }
 }
